move generation to service
