+++ Sources/General/Commands/Generate.swift
@@ -4,26 +4,11 @@
 
 import Foundation
 import ArgumentParser
-import Stencil
-import StencilSwiftKit
-import Yams
-import PathKit
 import GeneralKit
 
 public final class Generate: ParsableCommand {
-
-    enum Error: Swift.Error {
-        case noOutput(template: String)
-    }
-
-    private lazy var specFactory: SpecFactory = .init()
-    private lazy var fileManager: FileManager = .default
-
-    private lazy var generalSpec: GeneralSpec? = {
-        let pathURL = URL(fileURLWithPath: path, isDirectory: true)
-        let specURL = URL(fileURLWithPath: Constants.generalSpecName, relativeTo: pathURL)
-        return try? specFactory.makeSpec(url: specURL)
-    }()
+    
+    public static let configuration: CommandConfiguration = .init(commandName: "gen", abstract: "Generates modules from templates.")
 
     // MARK: - Lifecycle
 
@@ -31,8 +16,6 @@ public final class Generate: ParsableCommand {
     }
 
     // MARK: - Parameters
-
-    public static let configuration: CommandConfiguration = .init(commandName: "gen", abstract: "Generates modules from templates.")
 
     @Option(name: .shortAndLong, completion: .directory, help: "The path for the project.")
     var path: String = FileManager.default.currentDirectoryPath
@@ -49,123 +32,10 @@ public final class Generate: ParsableCommand {
     @Argument(help: "The additional variables for templates.")
     var variables: [GeneralKit.Variable] = []
 
-    private var context: [String: Any] {
-        let year = Calendar.current.component(.year, from: .init())
-        var context: [String: Any] = ["name": name,
-                                      "year": year]
-        for variable in variables {
-            context[variable.key] = variable.value
-        }
-        return context
-    }
-
     // MARK: - Lifecycle
 
     public func run() throws {
-        //create environment and spec
-        let templatesURL = defineTemplatesURL()
-        let templateURL = templatesURL + template
-        let specURL = templateURL + Constants.specFilename
-        let templateSpec: TemplateSpec = try specFactory.makeSpec(url: specURL)
-
-        let environment = try makeEnvironment(templatesURL: templatesURL, templateURL: templateURL)
-        try add(templateSpec, environment: environment)
-        print("🎉 \(template) template with \(name) name was successfully generated.")
-    }
-
-    // MARK: - Private
-
-    private func defineTemplatesURL() -> URL {
-        let folderName = Constants.templatesFolderName
-        let localPath = "./\(folderName)/"
-        if fileManager.fileExists(atPath: localPath + template) {
-            return URL(fileURLWithPath: localPath)
-        }
-        return fileManager.homeDirectoryForCurrentUser + folderName
-    }
-
-    private func makeEnvironment(templatesURL: URL, templateURL: URL) throws -> Environment {
-        let commonTemplatesURL = templatesURL + Constants.commonTemplatesFolderName
-        let contents = try fileManager.contentsOfDirectory(at: templateURL,
-                                                           includingPropertiesForKeys: [.isDirectoryKey],
-                                                           options: [])
-        let directoryPaths: [Path] = contents.compactMap { url in
-            guard url.hasDirectoryPath else {
-                return nil
-            }
-            return Path(url.path)
-        }
-        var paths = [Path(commonTemplatesURL.path), Path(templateURL.path)]
-        paths.append(contentsOf: directoryPaths)
-        let environment = Environment(loader: FileSystemLoader(paths: paths))
-        environment.extensions.forEach { ext in
-            ext.registerStencilSwiftExtensions()
-        }
-        return environment
-    }
-
-    private func add(_ templateSpec: TemplateSpec, environment: Environment, completion: ((URL) throws -> Void)? = nil) throws {
-        for file in templateSpec.files {
-            // render template for the file based on common and template files
-            let rendered = try environment.renderTemplate(name: file.template, context: context).trimmingCharacters(in: .whitespacesAndNewlines)
-            let module = name
-            let fileName = file.fileName(in: module)
-
-            // make output url for the file
-            var outputURL = URL(fileURLWithPath: path)
-            if let output = file.output {
-                outputURL.appendPathComponent(output)
-            }
-            else if let folder = outputFolder() {
-                outputURL.appendPathComponent(folder)
-                if let suffix = templateSpec.suffix {
-                    outputURL.appendPathComponent(name + suffix)
-                }
-                else {
-                    outputURL.appendPathComponent(name)
-                }
-                if let subfolder = file.folder {
-                    outputURL.appendPathComponent(subfolder)
-                }
-                outputURL.appendPathComponent(fileName)
-            }
-            else {
-                throw Error.noOutput(template: template)
-            }
-
-            // write rendered template to file
-            let fileURL = outputURL
-            outputURL.deleteLastPathComponent()
-            guard !fileManager.fileExists(atPath: fileURL.path) else {
-                print(yellow("File already exists: \(fileURL.path)"))
-                continue
-            }
-            try fileManager.createDirectory(at: outputURL, withIntermediateDirectories: true, attributes: nil)
-            try rendered.write(to: fileURL, atomically: true, encoding: .utf8)
-            try completion?(fileURL)
-        }
-    }
-
-    private func outputFolder() -> String? {
-        if let output = self.output {
-            return output
-        }
-        guard let generalSpec = generalSpec,
-            let output = generalSpec.output(forTemplateName: template) else {
-            return nil
-        }
-        return output.path
-    }
-}
-
-// MARK: - Generate.Error
-
-extension Generate.Error: CustomStringConvertible {
-
-    var description: String {
-        switch self {
-            case .noOutput(let template):
-                return "There is no output path for \(template) template. Please use --output option or add output to general.yml."
-        }
+        let renderer = Renderer(name: name, template: template, path: path, variables: variables, output: output)
+        try renderer.render()
     }
 }
